Initial app accepting required parameters
This commit adds the basic functionality of accepting parameters and performing
nontrivial actions with them. It uses the relation engine client to access
the Jacobson data (currently hard-coded to the CI instance), computes the most
common GO terms for each cluster containing the given gene_keys. This is
accomplished by checking out the Jacobson data and making in-memory sql queries
using sqlite and pandas. Linting and a basic updates to app metadata are
included.