Updated README.md and pre-commit hook.
This commit updates the contribution instructions, and updates the pre-commit
hook to cover new files. The pre-commit python linting is purposefully limited
due to the specific way that kb-sdk modifies its own python files. More work
needs to be done in the README to explain this situation and the solutions
adopted in this repo.