Added RWRtools.
This commit adds the requirements necessary to run the RWRtools scripts. These
are currently stored in a private repo, so they may not be cloned directly. The
workaround for this is to upload a recent copy of the repository into shock and
use that node as reference data. In the future, this should be avoidable by
cloning the public repository directly.