Added RWRtools RWR_LOE app.
This commit adds the RWRtools script RWR_LOE to the module as an app. This
allows users to run a context analysis by supplying seed genes alone or
focusing on a target gene set. The reference data is updated to fix a bug in
RWR_LOE. The visualization takes the highest ranking genes (and the target
genes if applicable) from the output of RWR_LOE and displays the network
connections between them and the seed nodes.