Ensure use of the correct object name for state.
This commit finalizes the state persistence functionality for reports in this
app. This works by creating a new workspace object when the report is
generated, and the workspace ref to this object is passed into the report. The
report loads the state on initial load and then updates this object with state
periodically as it is changed. This solution works well for an individual
narrative, however the state will not, at present, survive a copy. This problem
can be worked around by adding state export and import functionality. This will
allow users to move state from one report to another without having to manually
recover the state.