Refer to state object by objid instead of name.
Workspace object objids are preserved on copy, therefore it is safe to refer to
the object id, so long as the correct workspace can be divined. For local
development, a prompt is used to store a workspace id in localStorage. In an
environment, the report is served using HTMLFileSetServ which has a predictable
URL structure. This tightly couples this app to the current version of this URL
structure, so this app will need to be updated when the URL structure
HTMLFileSetServ changes.