Changed network query source to reference data.
Previous to this commit, the relation engine (RE) is where data resided and the
app would query RE to produce the output graph. Now the data is loaded directly
into the app's reference data (refdata). This makes it so that the underlying
network data can be updated without a deploy to RE. This should make updating
data easier.

The data load process happens at refdata update time, which is when the data
version is larger than any previous data version at registration time. This is
managed by the scripts/refdata-load.sh which gets all the required data from
the data sources (currently three repos). The data is validated according to
the RE validation process. The data is then loaded into a sqlite3 database
using the scripts/network_upload.py script. Heterogeneous data sources are
merged and made available in this way to the app. The new data is slightly
different than the data in RE, therefore some changes to the app had to be
made. For this update, the most important relationships are as follows. The
node_ids and MapMan data have a one-to-many relationship. Transcripts have a
one-to-one relationship with MapMan data, and a one-to-many relationship with
GO Terms. Finally, GO Terms have a many-to-many relationship with node_ids.
In this commit, a new Transcripts column is added do the report table and the
MapMan column has updated to support displaying multiple bins. The CSS changes
largely support this effort. Finally, new networks of edges have been added
which currently requires an update to edgeMetadata in the report style code.