Added new networks and edge metadata loading.
This commit loads the latest network and gene data from ORNL as of 2021-08-12.
It loads the manifest.yaml file from the exascale_data repository and the
edge_type.yaml specification file from the relation_engine repository and
merges them into a metadata file which is presented in the report output.

Some edges in the new networks have scores outside the (0,1) interval and
needed to be scaled differently.