Added multiplex selection.
This commit adds a choice of which set of multiplex networks to use via a
selection menu in the RWR CV and RWR LOE apps. This also adds tests which run
queries against each available multiplex network. There is some corruption in
at least one of the data files in this iteration, that is, the All Layers file.