Changed tests and refactored modules slightly.
Performing an end to end test has an amount of overhead which is amplified in
parametrized tests. These tests are verifying that the multiplex networks can
be queried correctly, which can be done locally. The updates to the tests run
each RWR tool using each multiplex network. This commit is also a good time to
partition the code describing the app and all the helper functions. Finally,
the tests are now ordered so that when the postinstall script is run the final
test output remains on disk so that the report can be viewed and developed.