Improved render and sort by separating concerns.
This commit improves render time by limiting the number of times cytoscape has
to use its layout functionality. This can be further improved with changes to
the nodeRepulsion function. It also decouples computing the table data from
displaying the table data. There is more work to be done in this area. These
changes solve a bug where a tippy would be frozen if it was shown during a
state change and generally improve the user experience.