Added apt upgrade to update certificates.
The base image contains stale root certificates which caused builds to fail.
This happened when requesting the node setup script which uses a Let's Encrypt
signed certificate. The site's certificate is up to date, but the old root
certificate in the container, which has long since been replaced, finally
expired. This was compounded initially because the script download failed
silently, but this updates it to fail at the first sign of trouble.