Add environment detection in build and tests.
This commit adds environment detection (based on the configured service URL) to
support one codebase on multiple environments. The detection in the build is
required at the moment since the RWR Tools are still loaded as refdata and they
need to be loaded from the correct blobstore. The detection in the tests could
be eliminated by creating the needed objects in the tests or by mocking the
results of the workspace query through DataFileUtil.