Added featureset input to RWR_LOE.
This commit adds featureset inputs to the RWR_LOE app for seeds and targets and
updates the tests with mocks for these featuresets. The remaining tests for the
RWR_CV app still use workspace objects, but these tests form a nice integration
test for the workspace.

The mock DFU client is ad hoc which is not ideal, but this helps provide a nice
self-contained solution that does the job. My impression is that it is the most
readable approach for now. This commit fixes this pattern so that it is defined
as a mock client class which is instantiated much like the other catch-all mock
client EchoMock.