Pinned KBase dependencies to specific commits.
Recent updates to the relation engine code force a python version update which
is not necessary for this project at this time. The simplest fix for now is to
pin the dependencies to a specific commit. This version and the version in the
exascale_data repository are now pinned to a specific commit.